Go Data Structures, version 4.
Bike-shedding the AVL tree API. Hopefully the last such API change.

Compared to version 3, the AVL tree API has minor changes. The Clear
method no longer accepts a release function as it had limitations in
that it was not allowed to fail. The FindLowest and FindHighest methods
are renamed to First and Last respectively.

The API direction is to provide generally useful operations and avoid
special cases for improved convenience in some situations, or for
slightly better performance.